+++ Alloy/Shaders/Shaders.metal
@@ -27,8 +27,8 @@ struct BlockSize {
 template <typename T>
 void textureCopy(texture2d<T, access::read> sourceTexture,
                  texture2d<T, access::write> destinationTexture,
-                 constant uint2& readOffset,
-                 constant uint2& writeOffset,
+                 constant ushort2& readOffset,
+                 constant ushort2& writeOffset,
                  const ushort2 position) {
     const ushort2 textureSize = ushort2(sourceTexture.get_width(),
                                         sourceTexture.get_height());
@@ -45,8 +45,8 @@ void textureCopy(texture2d<T, access::read> sourceTexture,
 #define outerArguments(T)                                        \
 (texture2d<T, access::read> sourceTexture [[ texture(0) ]],      \
 texture2d<T, access::write> destinationTexture [[ texture(1) ]], \
-constant uint2& readOffset [[ buffer(0) ]],                      \
-constant uint2& writeOffset [[ buffer(1) ]],                     \
+constant ushort2& readOffset [[ buffer(0) ]],                      \
+constant ushort2& writeOffset [[ buffer(1) ]],                     \
 const ushort2 position [[thread_position_in_grid]])              \
 
 #define innerArguments \
